requirements.txt
aws-cdk-lib==2.194.0
cdklabs.generative-ai-cdk-constructs==0.1.309
constructs>=10.0.0,<11.0.0
